move util tests to the correct package
