+++ packages/util/tsconfig.json
@@ -1,4 +1,8 @@
 {
-  "extends": "./tsconfig.build.json",
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "resolveJsonModule": true
+  },
   "include": ["src", "typings", "tests", "tools"]
 }
